requirements.txt
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
sqlparse==0.3.1
python-dateutil==1.5
djangorestframework==3.12.4
psycopg2-binary==2.9.1
configparser==5.0.2

#Django
#-------------------------------------
django==3.2.4
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-filter==2.4.0  # https://github.com/carltongibson/django-filter
django-extensions==3.1.3 # https://github.com/django-extensions/django-extensions
dj-database-url==0.5.0
ipython==7.23.1
drf-yasg==1.20.0
django-rest-swagger==2.2.0
drf-spectacular==0.17.2
djangorestframework-simplejwt==4.7.1
PyJWT==2.1.0
django-cors-headers==3.7.0